requirements.txt
aiohttp==1.0.5
async-timeout==2.0.1
chardet==3.0.4
discord.py==0.16.12
multidict==4.1.0
nltk==3.2.5
six==1.11.0
websockets==3.4
